+++ Scripts/VR/InputDevice.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace cpvrlab_vr_suite.Scripts.VR
+namespace cpvr_vr_suite.Scripts.VR
 {
     public class InputDevice : MonoBehaviour
     {