+++ Scripts/VR/ResetOriginOnSceneChange.cs
@@ -1,7 +1,7 @@
 using UnityEngine;
 using UnityEngine.SceneManagement;
 
-namespace cpvrlab_vr_suite.Scripts.VR
+namespace cpvr_vr_suite.Scripts.VR
 {
     public class ResetOriginOnSceneChange : MonoBehaviour
     {